+++ src/EntityGraphQL/Schema/GraphQLMutationAttribute.cs
@@ -6,6 +6,7 @@ namespace EntityGraphQL.Schema
     /// Marks the method in the class as a Mutation for EntityGraphQL to include in the Mutation Type.
     /// You need to add the mutation to the schema using <code>schema.AddMutationFrom(new MyClass());</code>
     /// </summary>
+    [AttributeUsage(AttributeTargets.Method)]
     public class GraphQLMutationAttribute : Attribute
     {
         public GraphQLMutationAttribute(string description = null)
@@ -21,5 +22,6 @@ namespace EntityGraphQL.Schema
     /// Allows EntityGraphQL to know which argument of the mutation needs to be
     /// populated with the mutation arguments from the query
     /// </summary>
-    public interface IMutationArguments { }
+    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Parameter)]
+    public class MutationArgumentsAttribute : Attribute { }
 }