+++ src/tests/EntityGraphQL.Tests/MutationTests.cs
@@ -183,12 +183,16 @@ namespace EntityGraphQL.Tests
     internal class PeopleMutations
     {
         [GraphQLMutation]
+
+        [MutationArguments]
         public Person AddPerson(PeopleMutationsArgs args)
         {
             return new Person { Name = string.IsNullOrEmpty(args.Name) ? "Default" : args.Name, Id = 555 };
         }
 
         [GraphQLMutation]
+
+        [MutationArguments]
         public Expression<Func<TestSchema, Person>> AddPersonNames(TestSchema db, PeopleMutationsArgs args)
         {
             db.People.Add(new Person { Id = 11, Name = args.Names[0], LastName = args.Names[1] });
@@ -196,6 +200,8 @@ namespace EntityGraphQL.Tests
         }
 
         [GraphQLMutation]
+
+        [MutationArguments]
         public Person AddPersonInput(PeopleMutationsArgs args)
         {
             return new Person { Name = args.NameInput.Name, LastName = args.NameInput.LastName };
@@ -216,7 +222,9 @@ namespace EntityGraphQL.Tests
         }
     }
 
-    internal class PeopleMutationsArgs : IMutationArguments
+
+    [MutationArguments]
+    internal class PeopleMutationsArgs
     {
         public string Name { get; set; }
         public List<string> Names { get; set; }